Extract YNotebookRoom subclass to house kernel connection methods
Per dlqqq's review on #248: kernel methods are purely additive and don't
rely on modifying existing YRoom methods, making a subclass the natural fit.

YNotebookRoom(YRoom) owns:
- connect_kernel / disconnect_kernel lifecycle
- _connect_client / _fetch_kernel_info helpers
- _execution_worker / _run_item / execute_cell
- _find_kernel_cell

YRoomManager creates YNotebookRoom for 'json:notebook:*' rooms and YRoom
for everything else (text files, chat, global awareness).

Type safety: session_manager and execution_handlers now isinstance-check
for YNotebookRoom before calling kernel methods, making it a runtime error
to accidentally attach a kernel to a text-file room.
