+++ jupyter_server_documents/rooms/__init__.py
@@ -1,3 +1,4 @@
 from .yroom_file_api import YRoomFileAPI
 from .yroom_manager import YRoomManager
 from .yroom import YRoom
+from .ynotebook_room import YNotebookRoom
